requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
gunicorn>=21.2
psycopg2-binary>=2.9
dj-database-url>=2.1
python-dotenv>=1.0
whitenoise>=6.6
